@@ -17,32 +17,32 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.17",
-    "@next/mdx": "^10.0.4",
+    "@mdx-js/react": "^2.1.3",
+    "@next/mdx": "^13.0.6",
     "bootstrap": "^3.4.1",
     "date-fns": "^2.16.1",
     "disqus-react": "^1.0.11",
     "feed": "^4.2.1",
     "lunr": "^2.3.9",
-    "next": "^10.0.0",
+    "next": "^13.0.6",
     "next-compose-plugins": "^2.2.1",
     "next-fonts": "^1.5.1",
     "next-ga": "^2.3.4",
     "ramda": "^0.27.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-external-link": "^1.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-external-link": "^2.0.0",
     "react-highlight": "^0.14.0",
     "react-image-lightbox": "^5.1.1",
     "react-tweet-embed": "^1.2.2"
   },
   "devDependencies": {
-    "@mdx-js/loader": "^1.6.22",
+    "@mdx-js/loader": "^2.1.5",
     "@types/lunr": "^2.3.3",
     "@types/node": "^14.14.16",
     "@types/ramda": "^0.27.34",
-    "@types/react": "17.0.30",
-    "nextjs-sitemap-generator": "^1.1.3",
-    "optimize-css-assets-webpack-plugin": "^5.0.4",
+    "@types/react": "18.0.26",
+    "nextjs-sitemap-generator": "^1.3.1",
     "react-lunr": "^1.0.0",
     "typescript": "^4.1.3"
   }
