@@ -3,9 +3,7 @@ import Link from 'next/link';
 import { LinkProps } from './Link';
 
 const InternalLink: FC<LinkProps> = ({ children, className, ...rest }) => (
-  <Link {...rest}>
-    <a className={className}>{children}</a>
-  </Link>
+  <Link {...rest} className={className}>{children}</Link>
 );
 
 export default InternalLink;
